+++ sphinx/pyowm.commons.rst
@@ -4,6 +4,22 @@ pyowm.commons package
 Submodules
 ----------
 
+pyowm.commons.databoxes module
+------------------------------
+
+.. automodule:: pyowm.commons.databoxes
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+pyowm.commons.enums module
+--------------------------
+
+.. automodule:: pyowm.commons.enums
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 pyowm.commons.frontlinkedlist module
 ------------------------------------
 
@@ -24,14 +40,6 @@ pyowm.commons.image module
 --------------------------
 
 .. automodule:: pyowm.commons.image
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-pyowm.commons.enums module
---------------------------
-
-.. automodule:: pyowm.commons.enums
     :members:
     :undoc-members:
     :show-inheritance:
